packages/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hushspec"
version = "0.1.0"
description = "Portable specification for AI agent security rules"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[tool.pytest.ini_options]
testpaths = ["tests"]
